+++ internal/network/clock.go
@@ -1,4 +1,4 @@
-package server
+package network
 
 type compareResult int
