+++ internal/network/network.go
@@ -1,4 +1,4 @@
-package server
+package network
 
 import (
 	"fmt"
@@ -10,7 +10,7 @@ import (
 
 var wg = sync.WaitGroup{}
 
-func StartServer(nodeAmount int) {
+func BuildNetwork(nodeAmount int) {
 	for i := range nodeAmount {
 		wg.Add(1)
 		go startNode(strconv.Itoa(i), strconv.Itoa(nodeAmount))
@@ -21,7 +21,7 @@ func StartServer(nodeAmount int) {
 }
 
 func startNode(nodeId string, nodeAmount string) {
-	cmd := exec.Command("go", "run", "github.com/21Philip/Auction/internal/server/create-node", nodeId, nodeAmount)
+	cmd := exec.Command("go", "run", "github.com/21Philip/Auction/internal/network/create-node", nodeId, nodeAmount)
 	cmd.Stdin = os.Stdin
 	cmd.Stdout = os.Stdout
 	cmd.Stderr = os.Stderr
