+++ internal/network/node.go
@@ -1,4 +1,4 @@
-package server
+package network
 
 import (
 	"context"
